+++ database/migrations/2022_02_28_120033_create_critical_qty_measurement_scales_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('critical_qty_measurement_scales', function (Blueprint $table) {
-            $table->unsignedTinyInteger('id', true);
+            $table->unsignedSmallInteger('id', true); // min:0|max:65,535
             $table->string('name');
             $table->timestamps();
         });
